pyproject.toml
[project]
name = "llm-abstract-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.3",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "tiktoken>=0.12.0",
    "umap-learn>=0.5.9.post2",
]

[dependency-groups]
dev = [
    "google-genai>=1.31.0",
    "ipython>=9.4.0",
    "openai>=1.101.0",
]
